ui/bug-report/pyproject.toml
[project]
name = "bug-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nicegui>=3.9.0",
    "pyyaml>=6.0.3",
]
